Fixed issue when redirected token contains named instance in servername (#1568)
